Move SmtpTransport to separate package
