+++ k9mail-library/src/main/java/com/fsck/k9/mail/TransportProvider.java
@@ -6,7 +6,7 @@ import android.content.Context;
 import com.fsck.k9.mail.oauth.OAuth2TokenProvider;
 import com.fsck.k9.mail.ssl.DefaultTrustedSocketFactory;
 import com.fsck.k9.mail.store.StoreConfig;
-import com.fsck.k9.mail.transport.SmtpTransport;
+import com.fsck.k9.mail.transport.smtp.SmtpTransport;
 import com.fsck.k9.mail.transport.WebDavTransport;
 
 public class TransportProvider {
